refactor(views): move redirects helper functions
keep generate_redirect_uri as-is, update name/usage

merge deauthorize_redirect into its caller, the logout view
